Add a `display_list::background` module
